Update component to support OCP4.8+ console route configuration
The component has new parameters `openshift_version` and `route` which
should be used to configure a custom console hostname.

On OCP4.8 the component creates a resource locker patch to add the
custom console hostname to the `ingress.config.openshift.io/cluster`
custom resource, as documented in [customizing the console].

The parameter `openshift_version` is used to provide users the option to
still configure the OpenShift console hostname via the
`console.operator.openshift.io/cluster` custom resource on OCP4.7 and
older.

[customizing the console]: https://docs.openshift.com/container-platform/4.8/web_console/customizing-the-web-console.html#customizing-the-console-route_customizing-web-console