Create cert-manager Certificates in namespace `openshift-console`
This feature requires Kyverno to be installed on the cluster. The
component checks whether component `kyverno` is enabled on the
cluster if you configure cert-manager Certificate resources for the
web console. If the kyverno component isn't enabled, the component
compilation produces an error.

The component only depends on Kyverno for configurations which use
cert-manager Certificates for a custom console route.

Fixes #21