Generate favicon Route if a custom logo is configured
When the OCP console is customized, the default favicon is not served,
which is an intentional design decision, as documented in
https://bugzilla.redhat.com/show_bug.cgi?id=1844883#c1.

However, that bugzilla provides a workaround to restore the default
favicon when the console is customized, cf.
https://bugzilla.redhat.com/show_bug.cgi?id=1844883#c3.

We implement this workaround in this commit if the console logo is
customized, and a custom hostname is configured, as we otherwise can't
configure `spec.hostname` for the additional route.

Ideally, we'd like to allow users to configure a custom favicon if
they're customizing the logo, but that's currently not possible.

Fixes #27