Add support for patch-operator-backed patches
We modify the resources returned by `rl.Patch()` to inject custom ArgoCD
sync-wave annotations. In order to ensure this functionality isn't lost
when migrating to patch-operator-backed patches, we adjust the component
to also inject the sync-wave annotations for resources with `kind: Patch`.

Note that we aren't migrating to importing `patch-operator.libsonnet`
directly to make this change backwards-compatible.