Update default console plugins for OpenShift 4.17
OpenShift 4.17 configures the `networking-console-plugin` by default. We
update the component to add that plugin to the default plugins for
OpenShift >= 4.17.

Additionally, to keep the code reasonable, we remove support for
OpenShift < 4.14 and deploy the `monitoring-plugin` unconditionally.

Note that we leave the `upgrade-notification` test case untouched as the
remaining OpenShift < 4.17 case.